requirements.txt
pytest==7.0.1
apache-airflow-providers-discord==3.7.0
apache-airflow-providers-slack==8.7.1
